refactor: LottoGenerator를 통한 로또번호 생성 추가

+++ src/main/java/lotto/domain/Money.java
@@ -16,7 +16,7 @@ public class Money {
         this.wonValue = wonValue;
     }
 
-    public Money multiply(int multiplyNumber) {
+    public Money multiply(long multiplyNumber) {
         return new Money(wonValue.multiply(BigDecimal.valueOf(multiplyNumber)));
     }
